api/requirements.txt
# Main
django==4.0.7
django-filter>=21.1,<22
djangorestframework>=3.13.1,<4
djangorestframework-bulk>=0.2.1,<1
django-cors-headers>=3.10.0,<4
# Important to use this instead of the original library
git+https://github.com/bcgov/django-oidc-rp.git@50e6fa143e61b04849b4c66beef078be0d7669de#egg=django-oidc-rp

# Database managment
psycopg2-binary>=2.7.3.1,<3
pynacl>=1.4

# Web Server
gunicorn>=19.7.1,<20
whitenoise>=4.1.2,<4.2

# Documentation
django_rest_swagger>=2.1.2,<3

# Testing
pytest-cov>=2.5.1,<3
pytest-django>=3.1.2,<4

# Util
requests>=2.18.4,<3
pillow>=8.1.0
numpy>=1.19.5
jsonschema>=3.2.0,<4
zeep>=4.0.0